2019-12-29 Dockerize All Done
Back-end docker was working on Dockerfile and not working on docker compose.
Because I have added volume in docker-compsoe file.
It means all folder content was mounted to docker-daemon container. and also node_modules
Sharp on my local is for windows 32-64 and on docker daemon it should be for linux-64 platform.
So i have those issues. I fixed when I remove volume.